Update HondaTrunk.java
The comments I added are meant to make the code easier to follow. I suggested adding explanations about what the parameters in the super constructor mean and how the evaluateWinCondition method works, especially how it decides if there are 3 matches, 2 matches, or no matches. I also pointed out the need to explain why partial matches result in a smaller payout and how the calculation for that works. I asked for more details on why the bet amount is subtracted when there's no match and why the full match payout uses the returnAmt. Finally, I recommended clarifying the behavior of the default case in the payout method. These comments aim to help anyone reading the code understand why and how things are calculated.