Update HumanPlayer.java
The comments I added explain that the getInstance() method uses a singleton pattern to make sure there’s only one HumanPlayer in the application. I also suggested briefly mentioning that the money property is a JavaFX property, which helps update both the user interface and the backend at the same time.